+++ packages/delegate/src/createRequest.ts
@@ -22,7 +22,6 @@ import {
   OperationTypeNode,
   SelectionNode,
   SelectionSetNode,
-  VariableDefinitionNode,
 } from 'graphql';
 import { ICreateRequest } from './types.js';
 
@@ -84,8 +83,10 @@ export function createRequest({
     );
   }
 
-  const newVariables = Object.create(null);
-  const variableDefinitions: VariableDefinitionNode[] = [];
+  const newVariables = info?.variableValues ? { ...info.variableValues } : {};
+  const variableDefinitions = info?.operation.variableDefinitions
+    ? [...info.operation.variableDefinitions]
+    : [];
   const argNodes: ArgumentNode[] = [];
 
   if (args != null) {
@@ -104,18 +105,11 @@ export function createRequest({
       // Check if we can re-use the variable from the original request for this argument
       if (existingArgNode?.value.kind === Kind.VARIABLE) {
         const varName = existingArgNode.value.name.value;
-        const varValue = info?.variableValues?.[varName];
+        const varValue = newVariables[varName];
         // If the variable value is the same as the argument value,
         // we can re-use the variable and its definition
         if (varValue === argValue) {
           argNodes.push(existingArgNode);
-          const varDef = info?.operation.variableDefinitions?.find(
-            (varDef) => varDef.variable.name.value === varName,
-          );
-          if (varDef) {
-            variableDefinitions.push(varDef);
-          }
-          newVariables[varName] = varValue;
           continue;
         }
       }
@@ -248,11 +242,7 @@ function projectArgumentValue(argValue: any, argType: GraphQLInputType): any {
       projectArgumentValue(item, argType.ofType),
     );
   }
-  if (
-    isInputObjectType(argType) &&
-    typeof argValue === 'object' &&
-    argValue !== null
-  ) {
+  if (isInputObjectType(argType) && typeof argValue === 'object') {
     const projectedValue: any = {};
     const fields = argType.getFields();
     for (const key in argValue) {
@@ -266,16 +256,14 @@ function projectArgumentValue(argValue: any, argType: GraphQLInputType): any {
     }
     return projectedValue;
   }
-  if (argValue != null) {
-    if (argType.name === 'Boolean') {
-      return Boolean(argValue);
-    }
-    if (argType.name === 'Int' || argType.name === 'Float') {
-      return Number(argValue);
-    }
-    if (argType.name === 'String') {
-      return String(argValue);
-    }
+  if (argType.name === 'Boolean') {
+    return Boolean(argValue);
+  }
+  if (argType.name === 'Int' || argType.name === 'Float') {
+    return Number(argValue);
+  }
+  if (argType.name === 'String') {
+    return String(argValue);
   }
   return argValue;
 }