+++ packages/fusion-runtime/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-mesh/fusion-runtime
 
+## 1.6.10
+### Patch Changes
+
+- Updated dependencies [[`c75fd0a`](https://github.com/graphql-hive/gateway/commit/c75fd0a64730ab07a3a7115748841025e312f335)]:
+  - @graphql-hive/logger@1.1.0
+  - @graphql-mesh/transport-common@1.0.15
+
+## 1.6.9
+### Patch Changes
+
+- Updated dependencies [[`4065b7f`](https://github.com/graphql-hive/gateway/commit/4065b7fbb08d9e75c5f0d3b2b4d42d665aa9dbd9)]:
+  - @graphql-tools/delegate@12.0.8
+  - @graphql-tools/federation@4.2.12
+  - @graphql-tools/stitch@10.1.12
+  - @graphql-tools/stitching-directives@4.0.14
+  - @graphql-tools/wrap@11.1.8
+
 ## 1.6.8
 ### Patch Changes