+++ packages/plugins/aws-sigv4/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-hive/plugin-aws-sigv4
 
+## 2.0.28
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-mesh/fusion-runtime@1.6.10
+
+## 2.0.27
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-mesh/fusion-runtime@1.6.9
+
 ## 2.0.26
 ### Patch Changes