+++ packages/router-runtime/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @graphql-hive/router-runtime
 
+## 1.1.13
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-mesh/fusion-runtime@1.6.10
+  - @graphql-mesh/transport-common@1.0.15
+
+## 1.1.12
+### Patch Changes
+
+
+
+- [#2001](https://github.com/graphql-hive/gateway/pull/2001) [`0408878`](https://github.com/graphql-hive/gateway/commit/04088789b84c037886c88c6545fa2452c55927cc) Thanks [@ardatan](https://github.com/ardatan)! - Fix projection while using Rust QP
+
+- Updated dependencies []:
+  - @graphql-mesh/fusion-runtime@1.6.9
+
+## 1.1.11
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-tools/federation@4.2.12
+  - @graphql-mesh/fusion-runtime@1.6.9
+
 ## 1.1.10
 ### Patch Changes